+++ InputKit/InputKit.csproj
@@ -1,54 +1,35 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <!--<TargetFrameworks Condition=" '$(OS)' == 'Windows_NT' ">netstandard2.0;netstandard1.0;MonoAndroid90;MonoAndroid10.0;MonoAndroid11.0;Xamarin.iOS10;uap10.0.19041</TargetFrameworks>
-    <TargetFrameworks Condition=" '$(OS)' != 'Windows_NT' ">netstandard2.0;netstandard1.0;MonoAndroid90;MonoAndroid10.0;MonoAndroid11.0;Xamarin.iOS10</TargetFrameworks>-->
-    <TargetFrameworks>net6.0-android;net6.0-ios;net6.0-maccatalyst</TargetFrameworks>
 
-    <AssemblyName>Plugin.InputKit</AssemblyName>
-    <RootNamespace>Plugin.InputKit</RootNamespace>
-    <PackageId>Xamarin.Forms.InputKit</PackageId>
-    <Product>$(AssemblyName) ($(TargetFramework))</Product>
-    <Version>3.7.2</Version>
-    <PackageVersion>3.7.2</PackageVersion>
-    <PackOnBuild>false</PackOnBuild>
-    <NeutralLanguage>en-US</NeutralLanguage>
-    <DefineConstants>$(DefineConstants);</DefineConstants>
-    <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
-    <GenerateDocumentationFile Condition=" '$(Configuration)' == 'Release' ">true</GenerateDocumentationFile>
-    <UseFullSemVerForNuGet>false</UseFullSemVerForNuGet>
-    <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
+    <!--<TargetFrameworks>netstandard2.0;net6.0-android;net6.0-ios;net6.0-maccatalyst</TargetFrameworks>-->
+    <TargetFrameworks>netstandard2.0;net6.0-android;net6.0-ios</TargetFrameworks>
+    <LangVersion>latest</LangVersion>
 
-    <!--TODO: Fill these in-->
-    <PackageLicenseUrl></PackageLicenseUrl>
-    <PackageProjectUrl>https://github.com/enisn/Xamarin.Forms.InputKit</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/enisn/Xamarin.Forms.InputKit</RepositoryUrl>
-    <PackageReleaseNotes>Android Target Framework issue resolved.</PackageReleaseNotes>
-    <PackageTags>xamarin, windows, ios, android, xamarin.forms, plugin, InputKit, checkbox, radiobutton,radio button, slider, dropdown</PackageTags>
-
-    <Title>Common controls mostly required</Title>
-    <Summary>Checkbox, Radio Button, Advanced Slider included this package </Summary>
-    <Description>Ready to use Input componenets such as Checkbox,  RadioButton, Dropdown, AdvancedEntry, Advanced Slider etc.</Description>
-
-    <Owners>enisn</Owners>
-    <Authors>enisn</Authors>
-    <Copyright>Copyright 2018</Copyright>
-    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <SignAssembly>false</SignAssembly>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageIcon>icon.png</PackageIcon>
   </PropertyGroup>
 
   <ItemGroup>
-    <!-- <PackageReference Include="Microsoft.Maui.Controls.Compatibility" Version="4.8.0.1364" PrivateAssets="All" /> -->
-    <PackageReference Include="Microsoft.Maui.Controls.Compatibility" Version="6.0.100-preview.5.794" />
-    <Compile Include="Shared\**\*.cs" />
+    <PackageReference Include="Microsoft.Maui.Controls.Compatibility" Version="6.0.100-preview.5.794" PrivateAssets="All" />
+    <!--<Compile Include="Shared\**\*.cs" />-->
   </ItemGroup>
 
-  <ItemGroup Condition="$(TargetFramework.Contains('-android'))">
-
-    <Compile Include="Platforms\Droid\**\*.cs" />
-
+  <ItemGroup Condition="$(TargetFramework.StartsWith('Xamarin.iOS')) != true AND $(TargetFramework.StartsWith('net6.0-ios')) != true AND $(TargetFramework.StartsWith('net6.0-maccatalyst')) != true ">
+    <Compile Remove="**\**\*.iOS.cs" />
+    <None Include="**\**\*.iOS.cs" />
+    <Compile Remove="**\iOS\**\*.cs" />
+    <None Include="**\iOS\**\*.cs" />
+  </ItemGroup>
+  <!--<ItemGroup Condition="$(TargetFramework.StartsWith('Xamarin.Mac')) != true ">
+    <Compile Remove="**\*.Mac.cs" />
+    <None Include="**\*.Mac.cs" />
+    <Compile Remove="**\Mac\**\*.cs" />
+    <None Include="**\Mac\**\*.cs" />
+  </ItemGroup>-->
+  <ItemGroup Condition="$(TargetFramework.StartsWith('MonoAndroid')) != true AND $(TargetFramework.StartsWith('net6.0-android')) != true ">
+    <Compile Remove="**\**\*.Android.cs" />
+    <None Include="**\**\*.Android.cs" />
+    <Compile Remove="**\Android\**\*.cs" />
+    <None Include="**\Android\**\*.cs" />
     <AndroidResource Include="Resources\drawable\*.xml" />
     <AndroidResource Include="Resources\drawable\*.png" />
     <AndroidResource Include="Resources\drawable-hdpi\*.png" />
@@ -57,56 +38,40 @@
     <AndroidResource Include="Resources\drawable-xxxhdpi\*.png" />
     <AndroidResource Include="Resources\values\*.xml" />
   </ItemGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0-ios' ">
-    <Compile Include="Platforms\iOS\**\*.cs" />
-
-    <!--<Copy SourceFiles="$(NuSpecPath)" DestinationFolder="$(PackageSource)\Resources"/>-->
-
-    <!--<Content Include="Resources\*.png">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>-->
+  <ItemGroup Condition="$(TargetFramework.StartsWith('netstandard')) != true ">
+    <Compile Remove="**\*.Standard.cs" />
+    <None Include="**\*.Standard.cs" />
+    <Compile Remove="**\Standard\**\*.cs" />
+    <None Include="**\Standard\**\*.cs" />
   </ItemGroup>
-
+  <!--<ItemGroup Condition="$(TargetFramework.Contains('-windows')) != true ">
+    <Compile Remove="**\*.Windows.cs" />
+    <None Include="**\*.Windows.cs" />
+    <Compile Remove="**\Windows\**\*.cs" />
+    <None Include="**\Windows\**\*.cs" />
+    <Compile Remove="**\*.uwp.cs" />
+    <None Include="**\*.uwp.cs" />
+    <MauiXaml Remove="**\*.Windows.xaml" />
+    <None Include="**\*.Windows.xaml" />
+    <Compile Remove="**\*.Windows.xaml.cs" />
+    <None Include="**\*.Windows.xaml.cs" />
+    <MauiXaml Remove="**\Windows\**\*.xaml" />
+    <None Include="**\Windows\**\*.xaml" />
+    <Compile Remove="**\Windows\**\*.xaml.cs" />
+    <None Include="**\Windows\**\*.xaml.cs" />
+  </ItemGroup>
   <ItemGroup Condition=" $(TargetFramework.StartsWith('uap10.0')) ">
-    <PackageReference Include="Microsoft.NETCore.UniversalWindowsPlatform" Version="6.0.6" />
-    <Compile Include="Platforms\UWP\**\*.cs" />
-    <!--
-    If you need to add sdk references:
-     <SDKReference Include="WindowsMobile, Version=10.0.16299.0">
-      <Name>Windows Mobile Extensions for the UWP</Name>
-    </SDKReference>
-    -->
+    <Compile Remove="**\*.uwp.cs" />
+    <None Include="**\*.uwp.cs" />
+  </ItemGroup>-->
+  <ItemGroup>
+    <Compile Remove="bin\**;obj\**" />
+    <None Remove="bin\**;obj\**" />
   </ItemGroup>
-
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'Xamarin.TVOS10' ">
-    <Compile Include="Platforms\TVOS\**\*.cs" />
-  </ItemGroup>
-
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'Xamarin.WatchOS10' ">
-    <Compile Include="Platforms\WatchOS\**\*.cs" />
-  </ItemGroup>
-
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'Xamarin.Mac20' ">
-    <Compile Include="Platforms\macOS\**\*.cs" />
-  </ItemGroup>
-
 
   <ItemGroup>
     <EmbeddedResource Include="Shared\Resources\**\*.png" />
     <EmbeddedResource Include="Shared\Resources\**\*.jpg" />
-  </ItemGroup>
-
-  <ItemGroup Condition="$(TargetFramework.Contains('-android'))">
-    <Compile Include="Platforms\Droid\AutoCompleteViewRenderer.cs" />
-    <Compile Include="Platforms\Droid\Config.cs" />
-    <Compile Include="Platforms\Droid\EmptyEntryRenderer.cs" />
-    <Compile Include="Platforms\Droid\MenuEffect.cs" />
-    <Compile Include="Platforms\Droid\NewIconViewRenderer.cs" />
-    <Compile Include="Platforms\Droid\StatefulStackLayoutRenderer.cs" />
   </ItemGroup>
 
   <ItemGroup>
@@ -135,9 +100,4 @@
     <Compile Include="Platforms\DotNet\**\*.cs" />
   </ItemGroup>
   -->
-
-
-
-  <Import Project="inputkit.targets" Condition="Exists('inputkit.targets')" />
-  <Import Project="$(MSBuildSDKExtrasTargets)" Condition="Exists('$(MSBuildSDKExtrasTargets)')" />
 </Project>
