--- InputKit/Platforms/UWP/InputKitImplementation.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Plugin.InputKit
-{
-    /// <summary>
-    /// Interface for $safeprojectgroupname$
-    /// </summary>
-    public class InputKitImplementation : IInputKit
-    {
-    }
-}
